Refactor e2e setup and helpers to tests/fixture/e2e for reuse
